[FIX] Wrong ACL usage
If `fw_for` is false, it means the headers are already there, so we
should check for the X-Forwarded-Host header instead of the Host one.

OTOH, if it is true, it means no X-Forwarded-* headers are present, so
we should check the Host header to know if we need to redirect.

Current rules were the opposite.